requirements.txt
discord.py==2.3.1
grequests==0.7.0
Requests==2.31.0
schedule==1.2.0